requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
django==3.2
django-filter==2.4.0
djangorestframework
djangorestframework-simplejwt==4.6.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
django-cors-headers


gunicorn==20.1.0
psycopg2-binary==2.8.6
factory_boy
Faker